+++ framework/db-local/src/main/java/module-info.java
@@ -30,7 +30,7 @@ module is.codion.framework.db.local {
 	requires transitive is.codion.framework.db.core;
 
 	exports is.codion.framework.db.local;
-	exports is.codion.framework.db.local.logger
+	exports is.codion.framework.db.local.tracer
 					to is.codion.framework.server;
 
 	provides is.codion.framework.db.EntityConnectionProvider.Builder