+++ changelog.md
@@ -15,6 +15,9 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - FilterTableCellRenderer.Builder.ColumnClassBuilder added.
 - FilterTable.Builder.rowHeight(), rowMargin(), intercellSpacing(), gridColor(), showGrid(), showHorizontalLines(), showVerticalLines() and dragEnabled() added.
+### is.codion.framework.db.local
+- MethodLogger renamed MethodTracer, logger package renamed tracer.
+- DefaultLocalEntityConnection.NoOpTracer added, tracer no longer nullable.
 ### is.codion.swing.framework.model
 - EntityApplicationModel.USER configuration value added.
 - EntityApplicationModel.USERNAME_PREFIX configuration value removed, not really used anymore.